feat: Added default IP Address
A new functionality was added to the PC version whereby the last IP address filled in the input field is suggested back to the user so as to avoid having to retype the IP address again if it hasn't changed. This way, provided the IP address of the phone remains the same, the user just has to click join room directly.